feat: split SqlObserver into operation-specific observers
Phase 2 complete:
- Extract shared utilities to SqlUtils class in src/lib/observers/
- Split monolithic SqlObserver into 5 operation-specific observers:
  * sql-create-observer.ts - handles CREATE operations
  * sql-update-observer.ts - handles UPDATE operations
  * sql-delete-observer.ts - handles DELETE operations
  * sql-select-observer.ts - handles SELECT operations
  * sql-revert-observer.ts - handles REVERT operations
- Each observer extends BaseObserver with focused single responsibility
- Clean execute() methods with direct business logic (no wrapper methods)
- Remove original monolithic SqlObserver

Benefits:
- Single Responsibility: Each observer handles exactly one operation
- Better Performance: Only relevant observers execute per operation
- Easier Testing: Individual observers simpler to unit test
- Maintainability: Changes isolated to specific operations
